requirements.txt
# Core dependencies
streamlit>=1.24.0
numpy>=1.24.0
pandas>=1.5.3
requests>=2.28.0
urllib3>=1.26.0

# Data handling and scheduling
pandas_market_calendars>=4.0.0
schedule>=1.2.0

# Visualization
plotly>=4.14.3
matplotlib>=3.7.0
seaborn>=0.12.0

# Web scraping
beautifulsoup4>=4.11.0

# nGS System dependencies (Polygon API)
polygon-api-client>=1.12.0
typing-extensions>=4.5.0

# Machine learning
scikit-learn>=1.3.0

# Rate limit handling
backoff>=2.2.1  # Added for exponential backoff retries

# Machine Learning (for AI integration manager)
scikit-learn>=1.2.0
